Move json files in root to data/json folder
